Moved ScanUtils to util.scans
