+++ src/main/java/net/sf/mzmine/datamodel/impl/SimpleFeature.java
@@ -30,7 +30,7 @@ import net.sf.mzmine.datamodel.IsotopePattern;
 import net.sf.mzmine.datamodel.RawDataFile;
 import net.sf.mzmine.util.PeakUtils;
 import net.sf.mzmine.util.RawDataFileUtils;
-import net.sf.mzmine.util.ScanUtils;
+import net.sf.mzmine.util.scans.ScanUtils;
 
 /**
  * This class is a simple implementation of the peak interface.
