+++ src/main/java/net/sf/mzmine/modules/masslistmethods/chromatogrambuilder/Chromatogram.java
@@ -33,7 +33,7 @@ import net.sf.mzmine.datamodel.Scan;
 import net.sf.mzmine.datamodel.impl.SimplePeakInformation;
 import net.sf.mzmine.main.MZmineCore;
 import net.sf.mzmine.util.MathUtils;
-import net.sf.mzmine.util.ScanUtils;
+import net.sf.mzmine.util.scans.ScanUtils;
 
 /**
  * Chromatogram implementing ChromatographicPeak.
