+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/io/casmiimport/CasmiImportTask.java
@@ -48,7 +48,7 @@ import net.sf.mzmine.main.MZmineCore;
 import net.sf.mzmine.parameters.ParameterSet;
 import net.sf.mzmine.taskcontrol.AbstractTask;
 import net.sf.mzmine.taskcontrol.TaskStatus;
-import net.sf.mzmine.util.ScanUtils;
+import net.sf.mzmine.util.scans.ScanUtils;
 
 class CasmiImportTask extends AbstractTask {
