+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/io/sqlexport/SQLExportTask.java
@@ -39,7 +39,7 @@ import net.sf.mzmine.parameters.ParameterSet;
 import net.sf.mzmine.taskcontrol.AbstractTask;
 import net.sf.mzmine.taskcontrol.TaskStatus;
 import net.sf.mzmine.util.RangeUtils;
-import net.sf.mzmine.util.ScanUtils;
+import net.sf.mzmine.util.scans.ScanUtils;
 
 class SQLExportTask extends AbstractTask {
