+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/peakpicking/peakextender/ExtendedPeak.java
@@ -15,7 +15,7 @@ import net.sf.mzmine.datamodel.impl.SimplePeakInformation;
 import net.sf.mzmine.main.MZmineCore;
 import net.sf.mzmine.util.MathUtils;
 import net.sf.mzmine.util.PeakUtils;
-import net.sf.mzmine.util.ScanUtils;
+import net.sf.mzmine.util.scans.ScanUtils;
 
 public class ExtendedPeak implements Feature {
   private SimplePeakInformation peakInfo;
