+++ src/main/java/net/sf/mzmine/modules/rawdatamethods/peakpicking/manual/ManualPeak.java
@@ -30,7 +30,7 @@ import net.sf.mzmine.datamodel.Scan;
 import net.sf.mzmine.datamodel.impl.SimplePeakInformation;
 import net.sf.mzmine.util.MathUtils;
 import net.sf.mzmine.util.PeakUtils;
-import net.sf.mzmine.util.ScanUtils;
+import net.sf.mzmine.util.scans.ScanUtils;
 
 /**
  * This class represents a manually picked chromatographic peak.
