+++ src/main/java/net/sf/mzmine/modules/rawdatamethods/peakpicking/msms/MsMsPeakPickingTask.java
@@ -34,7 +34,7 @@ import net.sf.mzmine.parameters.ParameterSet;
 import net.sf.mzmine.parameters.parametertypes.selectors.ScanSelection;
 import net.sf.mzmine.taskcontrol.AbstractTask;
 import net.sf.mzmine.taskcontrol.TaskStatus;
-import net.sf.mzmine.util.ScanUtils;
+import net.sf.mzmine.util.scans.ScanUtils;
 
 public class MsMsPeakPickingTask extends AbstractTask {
   private Logger logger = Logger.getLogger(this.getClass().getName());
