+++ src/main/java/net/sf/mzmine/modules/rawdatamethods/rawdataimport/fileformats/AgilentCsvReadTask.java
@@ -31,8 +31,7 @@ import net.sf.mzmine.datamodel.impl.SimpleScan;
 import net.sf.mzmine.project.impl.RawDataFileImpl;
 import net.sf.mzmine.taskcontrol.AbstractTask;
 import net.sf.mzmine.taskcontrol.TaskStatus;
-import net.sf.mzmine.util.ScanUtils;
-
+import net.sf.mzmine.util.scans.ScanUtils;
 import com.google.common.collect.Range;
 
 public class AgilentCsvReadTask extends AbstractTask {
