+++ src/main/java/net/sf/mzmine/modules/rawdatamethods/rawdataimport/fileformats/MzDataReadTask.java
@@ -44,7 +44,7 @@ import net.sf.mzmine.datamodel.impl.SimpleScan;
 import net.sf.mzmine.taskcontrol.AbstractTask;
 import net.sf.mzmine.taskcontrol.TaskStatus;
 import net.sf.mzmine.util.ExceptionUtils;
-import net.sf.mzmine.util.ScanUtils;
+import net.sf.mzmine.util.scans.ScanUtils;
 
 /**
  * This class read 1.04 and 1.05 MZDATA files.
