+++ src/main/java/net/sf/mzmine/modules/visualization/tic/TICDataSet.java
@@ -34,7 +34,7 @@ import net.sf.mzmine.main.MZmineCore;
 import net.sf.mzmine.taskcontrol.Task;
 import net.sf.mzmine.taskcontrol.TaskPriority;
 import net.sf.mzmine.taskcontrol.TaskStatus;
-import net.sf.mzmine.util.ScanUtils;
+import net.sf.mzmine.util.scans.ScanUtils;
 
 /**
  * TIC visualizer data set. One data set is created per file shown in this visualizer. We need to
