+++ src/main/java/net/sf/mzmine/modules/visualization/twod/TwoDVisualizerModule.java
@@ -33,8 +33,7 @@ import net.sf.mzmine.parameters.parametertypes.selectors.RawDataFilesSelectionTy
 import net.sf.mzmine.parameters.parametertypes.selectors.ScanSelection;
 import net.sf.mzmine.taskcontrol.Task;
 import net.sf.mzmine.util.ExitCode;
-import net.sf.mzmine.util.ScanUtils;
-
+import net.sf.mzmine.util.scans.ScanUtils;
 import com.google.common.collect.Range;
 
 /**
